Rename directory & update README a little.
