+++ include/dst/partial/dynamic_avg_segment_tree.hpp
