+++ include/dst/partial/dynamic_negate_segment_tree.hpp
