+++ test/boost_adaptive_pool.cpp
@@ -7,7 +7,7 @@
 #include <gtest/gtest.h>
 
 #include <boost/container/adaptive_pool.hpp>
-#include <dst/curried/dynamic_simple_get_set_segment_tree.hpp>
+#include <dst/partial/dynamic_simple_get_set_segment_tree.hpp>
 #include <random>
 
 #include "reference/seg_tree_reference_base.hpp"
