+++ test/test_simple.cpp
@@ -6,7 +6,7 @@
 
 #include <gtest/gtest.h>
 
-#include <dst/curried/dynamic_simple_get_set_segment_tree.hpp>
+#include <dst/partial/dynamic_simple_get_set_segment_tree.hpp>
 #include <random>
 
 #include "reference/seg_tree_reference_base.hpp"
